[CEN-1521] Declare numTrx as integer

+++ api/batch/src/main/java/it/gov/pagopa/rtd/transaction_filter/batch/model/AdeTransactionsAggregate.java
@@ -40,7 +40,7 @@ AdeTransactionsAggregate {
 
     @NotNull
     @NotBlank
-    short numTrx;
+    int numTrx;
 
     @NotNull
     @NotBlank
